refactor: remove legacy spinner
