+++ app/admin/stores/index.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 import RequireWallet from '../../../components/RequireWallet';
 
 const StoresScreen = React.lazy(() => import('./_StoresScreen'));
@@ -7,7 +7,7 @@ const StoresScreen = React.lazy(() => import('./_StoresScreen'));
 export default function AdminStoresRoute(props: any) {
   return (
     <RequireWallet>
-      <Suspense fallback={<Spinner label="Stores" />}>
+      <Suspense fallback={<Spinner />}>
         <StoresScreen {...props} />
       </Suspense>
     </RequireWallet>
