+++ app/product/[id].tsx
@@ -3,7 +3,7 @@ import { Text } from 'react-native';
 import { useLocalSearchParams } from 'expo-router';
 import { z } from 'zod';
 import { createValidateParams } from '@/lib/validateParams';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 
 const ProductScreen = React.lazy(() => import('./_ProductScreen'));
 const validateParams = createValidateParams(z.object({ id: z.string() }));
@@ -14,7 +14,7 @@ export default function ProductScreenRoute() {
     return <Text>Invalid product</Text>;
   }
   return (
-    <Suspense fallback={<Spinner label="Product" />}> 
+    <Suspense fallback={<Spinner />}> 
       <ProductScreen id={params.data.id} />
     </Suspense>
   );
