+++ app/reviews/index.tsx
@@ -18,7 +18,7 @@ import DatabaseService from '../../services/database';
 import { Product, Review, Order } from '../../types';
 import { useAuth } from '@/features/auth/AuthContext';
 import { useTheme } from '../../contexts/ThemeContext';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 import InfoModal from '../../components/InfoModal';
 import ConfirmationModal from '../../components/ConfirmationModal';
 import { useAuthModal } from '@/features/auth/AuthModalContext';
@@ -478,7 +478,7 @@ export default function ReviewsScreen() {
           <Text style={[styles.headerTitle, { color: colors.text.primary }]}>ביקורות</Text>
           <View style={commonStyles.spacer40} />
         </View>
-        <Spinner label="Loading reviews" />
+        <Spinner />
       </SafeAreaView>
     );
   }
