+++ app/store/[storeId]/admin/dashboard.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 import RequireWallet from '../../../../components/RequireWallet';
 
 const DashboardScreen = React.lazy(() => import('./_DashboardScreen'));
@@ -7,7 +7,7 @@ const DashboardScreen = React.lazy(() => import('./_DashboardScreen'));
 export default function DashboardRoute(props: any) {
   return (
     <RequireWallet>
-      <Suspense fallback={<Spinner label="Dashboard" />}>
+      <Suspense fallback={<Spinner />}>
         <DashboardScreen {...props} />
       </Suspense>
     </RequireWallet>
