+++ app/store/[storeId]/admin/deliveries.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 import RequireWallet from '../../../../components/RequireWallet';
 
 // Lazy-load the deliveries management screen to reduce the main bundle
@@ -12,7 +12,7 @@ const DeliveriesScreen = React.lazy(() =>
 export default function DeliveriesRoute(props: any) {
   return (
     <RequireWallet>
-      <Suspense fallback={<Spinner label="Deliveries" />}>
+      <Suspense fallback={<Spinner />}>
         <DeliveriesScreen {...props} />
       </Suspense>
     </RequireWallet>
