+++ app/store/[storeId]/admin/disputes.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 import RequireWallet from '../../../../components/RequireWallet';
 
 const DisputesScreen = React.lazy(() => import('./_DisputesScreen'));
@@ -7,7 +7,7 @@ const DisputesScreen = React.lazy(() => import('./_DisputesScreen'));
 export default function DisputesRoute(props: any) {
   return (
     <RequireWallet>
-      <Suspense fallback={<Spinner label="Disputes" />}>
+      <Suspense fallback={<Spinner />}>
         <DisputesScreen {...props} />
       </Suspense>
     </RequireWallet>
