+++ app/store/[storeId]/admin/orders.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 import RequireWallet from '../../../../components/RequireWallet';
 
 const OrdersScreen = React.lazy(() => import('./_OrdersScreen'));
@@ -7,7 +7,7 @@ const OrdersScreen = React.lazy(() => import('./_OrdersScreen'));
 export default function OrdersRoute(props: any) {
   return (
     <RequireWallet>
-      <Suspense fallback={<Spinner label="Orders" />}>
+      <Suspense fallback={<Spinner />}>
         <OrdersScreen {...props} />
       </Suspense>
     </RequireWallet>
