+++ app/store/[storeId]/admin/pricing-tiers.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 import RequireWallet from '../../../../components/RequireWallet';
 
 const PricingTiersScreen = React.lazy(() => import('./_PricingTiersScreen'));
@@ -7,7 +7,7 @@ const PricingTiersScreen = React.lazy(() => import('./_PricingTiersScreen'));
 export default function PricingTiersRoute(props: any) {
   return (
     <RequireWallet>
-      <Suspense fallback={<Spinner label="Pricing Tiers" />}>
+      <Suspense fallback={<Spinner />}>
         <PricingTiersScreen {...props} />
       </Suspense>
     </RequireWallet>
