+++ app/store/[storeId]/admin/products.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 import RequireWallet from '../../../../components/RequireWallet';
 
 const ProductsScreen = React.lazy(() => import('./_ProductsScreen'));
@@ -7,7 +7,7 @@ const ProductsScreen = React.lazy(() => import('./_ProductsScreen'));
 export default function ProductsRoute(props: any) {
   return (
     <RequireWallet>
-      <Suspense fallback={<Spinner label="Products" />}>
+      <Suspense fallback={<Spinner />}>
         <ProductsScreen {...props} />
       </Suspense>
     </RequireWallet>
