+++ app/store/[storeId]/admin/settings.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 import RequireWallet from '../../../../components/RequireWallet';
 
 const SettingsScreen = React.lazy(() => import('./_SettingsScreen'));
@@ -7,7 +7,7 @@ const SettingsScreen = React.lazy(() => import('./_SettingsScreen'));
 export default function SettingsRoute(props: any) {
   return (
     <RequireWallet>
-      <Suspense fallback={<Spinner label="Settings" />}>
+      <Suspense fallback={<Spinner />}>
         <SettingsScreen {...props} />
       </Suspense>
     </RequireWallet>
