+++ app/store/[storeId]/admin/user-management.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 import RequireWallet from '../../../../components/RequireWallet';
 
 const UserManagementScreen = React.lazy(() => import('./_UserManagementScreen'));
@@ -7,7 +7,7 @@ const UserManagementScreen = React.lazy(() => import('./_UserManagementScreen'))
 export default function UserManagementRoute(props: any) {
   return (
     <RequireWallet>
-      <Suspense fallback={<Spinner label="User Management" />}>
+      <Suspense fallback={<Spinner />}>
         <UserManagementScreen {...props} />
       </Suspense>
     </RequireWallet>
