+++ app/store/[storeId]/index.tsx
@@ -1,11 +1,11 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 
 const StoreScreen = React.lazy(() => import('./_StoreScreen'));
 
 export default function StoreRoute(props: any) {
   return (
-    <Suspense fallback={<Spinner label="Storefront" />}>
+    <Suspense fallback={<Spinner />}>
       <StoreScreen {...props} />
     </Suspense>
   );
