+++ app/storefront/index.tsx
@@ -1,11 +1,11 @@
 import React, { Suspense } from 'react';
-import Spinner from '@/shared/ui/Spinner';
+import { Spinner } from '@/ui/primitives';
 
 const StorefrontScreen = React.lazy(() => import('./_StorefrontScreen'));
 
 export default function StorefrontRoute(props: any) {
   return (
-    <Suspense fallback={<Spinner label="Storefront" />}>
+    <Suspense fallback={<Spinner />}>
       <StorefrontScreen {...props} />
     </Suspense>
   );
